requirements.txt
asgiref==3.5.2
Babel==2.10.3
dj-database-url==0.5.0
Django==4.0.5
django-phonenumber-field==6.3.0
gunicorn==20.1.0
phonenumbers==8.12.51
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
